refactor temp_dir to conftest
